Refactor(DocumentsCRUD)
Updated to current view.
Some fixes.
