+++ ClientApp/src/app/components/indicator-detail/file-document-form/file-document-form.component.spec.ts
