+++ ClientApp/src/app/components/indicator-detail/indicator-detail-registry/indicator-detail-registry.component.html
@@ -11,7 +11,7 @@
           <div fxFlex="42%">
             <div class="value-box" *ngIf="registry.discriminator === 'QuantityRegistry'">{{registry.quantity}}</div>
             <div class="value-box" *ngIf="registry.discriminator === 'PercentRegistry' ">{{registry.percent}}%</div>
-          </div>          
+          </div>       
           <div fxFlex="15%">
             <button class="btn btn-xs btn-primary btn-size-text" (click)="openModalEditRegistry($event, editRegistry,registry)">Modificar</button>
           </div>
@@ -22,6 +22,12 @@
         </div>
       </ng-template>
       <ng-template ngbPanelContent>
+        <div class="row">
+          <button class="btn btn-xs btn-primary btn-size-text pull-right" (click)="openModalFileDocument($event, fileModal, registry)">Agregar archivo</button>
+          <span class="pull-right">&nbsp;</span>
+          <button class="btn btn-xs btn-primary btn-size-text pull-right" (click)="openModalLinkDocument($event, linkModal, registry)">Agregar link</button>
+        </div>
+        <br/>
         <ng-template ngFor let-document [ngForOf]="registry.documents" let-i="index">
           <div fxLayout="row">
             <div fxFlex="40%">
@@ -37,10 +43,26 @@
               <br /> {{document.extension}}
             </div>
             <div fxFlex="10%">
-              <a class="btn btn-xs btn-primary">Ver</a>
+              <button *ngIf="document.extension=='link'; else viewFileTemplate" 
+              type="button" (click)="goToLinkBlank(document.link)" class="btn btn-xs btn-primary">
+                Ver
+              </button>
+              <ng-template #viewFileTemplate>
+                  <button type="button" class="btn btn-xs btn-primary disabled">
+                    Ver
+                  </button>
+              </ng-template>
             </div>
             <div fxFlex="15%">
-              <a class="btn btn-xs btn-info btn-size-text">Descargar</a>
+                <button *ngIf="document.extension=='link'; else downloadFileTemplate" 
+                type="button" (click)="goToLink(document.link)" class="btn btn-xs btn-info btn-size-text">
+                  Descargar
+                </button>
+                <ng-template #downloadFileTemplate>
+                    <button type="button" (click)="download(document)"  class="btn btn-xs btn-info btn-size-text">
+                      Descargar
+                    </button>
+                </ng-template>
             </div>
             <div fxFlex="10%">
               <a class="btn btn-xs btn-danger btn-size-text" (click)="deleteDocument(registry,document)">Eliminar</a>
@@ -58,3 +80,29 @@
 <ng-template #editRegistry>
   <app-registry-editor [registry]="registry" [registriesType]="registriesType" [editModalRef]="editModalRef" #editRegistry></app-registry-editor>
 </ng-template>
+
+<!-- Modals para agregar documentos -->
+<ng-template #linkModal>
+  <div class="modal-header">
+    <h4 class="modal-title pull-left">Ingrese un link externo</h4>
+    <button type="button" class="close pull-right" aria-label="Close" (click)="modalRef.hide()">
+      <span aria-hidden="true">&times;</span>
+    </button>
+  </div>
+  <div class="modal-body">
+    <app-link-document-form [registry]="registry" [idRegistry]="registry.registryID" [modalRef]="modalRef"></app-link-document-form>
+  </div>
+</ng-template>
+
+<ng-template #fileModal>
+  <div class="modal-header">
+    <h4 class="modal-title pull-left">Ingrese un documento de respaldo</h4>
+    <button type="button" class="close pull-right" aria-label="Close" (click)="modalRef.hide()">
+      <span aria-hidden="true">&times;</span>
+    </button>
+  </div>
+  <div class="modal-body">
+    <app-file-document-form [registry]="registry" [idRegistry]="registry.registryID" [modalRef]="modalRef"></app-file-document-form>
+  </div>
+</ng-template>
+<!-- FIN Modals para agregar documentos-->
