+++ ClientApp/src/app/components/indicator-detail/link-document-form/link-document-form.component.html
