--- ClientApp/src/app/components/registry-details/document-editor/document-editor.component.css
@@ -1,7 +1,0 @@
-.ng-valid[required], .ng-valid.required {
-  border-left: 5px solid #5cb85c; /* green */
-}
-
-.ng-invalid:not(form) {
-  border-left: 5px solid #d9534f; /* red */
-}
