--- ClientApp/src/app/components/registry-details/document-editor/document-editor.component.spec.ts
@@ -1,25 +1,0 @@
-import { async, ComponentFixture, TestBed } from '@angular/core/testing';
-
-import { DocumentEditorComponent } from './document-editor.component';
-
-describe('DocumentEditorComponent', () => {
-  let component: DocumentEditorComponent;
-  let fixture: ComponentFixture<DocumentEditorComponent>;
-
-  beforeEach(async(() => {
-    TestBed.configureTestingModule({
-      declarations: [ DocumentEditorComponent ]
-    })
-    .compileComponents();
-  }));
-
-  beforeEach(() => {
-    fixture = TestBed.createComponent(DocumentEditorComponent);
-    component = fixture.componentInstance;
-    fixture.detectChanges();
-  });
-
-  it('should create', () => {
-    expect(component).toBeTruthy();
-  });
-});
