--- ClientApp/src/app/components/registry-details/registry-details.component.html
@@ -1,93 +1,0 @@
-<section>
-    <div id="head-page-section" class="l-grey-bg head-page-section">
-      <div class="container">
-          <h1>    Nombre registro: {{ registry.name }}</h1>
-          <div style="float:right">
-            <button *ngIf="registry.discriminator == 'LinkRegistry'"
-                    type="button" (click)="openModal(linkModal)" class="btn btn-success">
-              Agregar desde link
-            </button>
-
-              <button type="button" (click)="openModal(fileModal)" class="btn btn-success">Agregar desde archivo</button>
-
-          </div>
-      </div>
-    </div>
-</section>
-
-<div class="container">
-  <table class="table" *ngIf="registry.documents.length != 0">
-    <thead>
-      <tr>
-        <td>Nombre</td>
-        <td>Documento</td>
-        <td>Extensión</td>
-        <td>Fecha</td>
-        <td>Código</td>
-        <td>Descargar</td>
-        <td>Modificar</td>
-        <td>Eliminar</td>
-      </tr>
-    </thead>
-    <tbody>
-      <tr *ngFor="let document of registry.documents">
-        <td>{{document.name}}</td>
-        <td>{{document.documentName}}</td>
-        <td>{{document.extension}}</td>
-        <td>{{document.date}}</td>
-        <td>{{document.code}}</td>
-        <td>
-          <button *ngIf="document.extension=='link'; else fileTemplate" 
-          type="button" (click)="goToLink(document.link)" class="btn btn-success">
-            <i class="glyphicon glyphicon-download-alt"></i>
-          </button>
-        </td>
-        <ng-template #fileTemplate>
-            <button type="button" (click)="download(document)" class="btn btn-success">
-              <i class="glyphicon glyphicon-download-alt"></i>
-            </button>
-        </ng-template>
-        <td>
-          <div class="pull-right">
-            <button href="#" class="btn btn-success btn" (click)="openModalEditDocument(editDocument, document)"> <span class="glyphicon glyphicon-pencil">  </span>  </button></div>
-        </td>
-        <td>
-          <div class="pull-right">
-           <button href="#" class="btn btn-danger btn" (click)="deleteDocument(registry,document)"> <span class="glyphicon glyphicon-trash"> </span> </button>
-          </div>
-        </td>
-      </tr>
-    </tbody>
-  </table>
-
-  <p *ngIf="registry.documents">No Hay elementos para mostrar</p>
-</div>
-<br>
-
-<ng-template #linkModal>
-  <div class="modal-header">
-    <h4 class="modal-title pull-left">Ingrese un link externo</h4>
-    <button type="button" class="close pull-right" aria-label="Close" (click)="modalRef.hide()">
-      <span aria-hidden="true">&times;</span>
-    </button>
-  </div>
-  <div class="modal-body">
-    <app-link-document-form [registry]="registry" [idRegistry]="idRegistry" [modalRef]="modalRef"></app-link-document-form>
-  </div>
-</ng-template>
-
-<ng-template #fileModal>
-  <div class="modal-header">
-    <h4 class="modal-title pull-left">Ingrese un documento de respaldo</h4>
-    <button type="button" class="close pull-right" aria-label="Close" (click)="modalRef.hide()">
-      <span aria-hidden="true">&times;</span>
-    </button>
-  </div>
-  <div class="modal-body">
-    <app-file-document-form [registry]="registry" [idRegistry]="idRegistry" [modalRef]="modalRef"></app-file-document-form>
-  </div>
-</ng-template>
-
-<ng-template #editDocument>
-  <app-document-editor [document]="document" [discriminator]="discriminator" [editModalRef]="editModalRef" #editDocument></app-document-editor>
-</ng-template>
