+++ ClientApp/src/app/services/registry/registry.service.ts
@@ -25,14 +25,9 @@ export class RegistryService {
       return this.http.get<Registry>(RegistryService.REGISTRIES_API + registryId);
   }
 
-
-  addLinkDocument(document: Document, registryId: number) {
-    this.http.post<Registry>(RegistryService.REGISTRIES_API + registryId
-      + RegistryService.ADD_LINK_DOCUMENT_METHOD, document ).subscribe();
-  }
-  addFileDocument(document: Document, registryId: number) {
-    this.http.post<Registry>(RegistryService.REGISTRIES_API + registryId
-      + RegistryService.ADD_FILE_DOCUMENT_METHOD, document ).subscribe();
+  addLinkDocument(document: Document, registryId: number): Observable<Document> {
+    return this.http.post<Document>(RegistryService.REGISTRIES_API + registryId
+      + RegistryService.ADD_LINK_DOCUMENT_METHOD, document );
   }
 
   editRegistry(registry: Registry, registriesType: number): Observable<Registry> {
